+++ NEWS.md
@@ -1,4 +1,4 @@
-# Development
+# 0.2.0
 
 New: